jzftdz.com/jzftdz_scraper/requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.1
PyYAML==6.0.1
requests==2.31.0
